requirements.txt
Flask>=3.0.0
wx-crypt>=0.0.2
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
httpx>=0.27.0
pydantic~=2.13.3
pycryptodome~=3.23.0